spring-data

+++ src/main/java/com/igormulyar/lcts/controller/CourtCasesRepo.java
@@ -10,7 +10,7 @@ import java.util.List;
  * Created by igor on 07.01.17.
  */
 @Component
-public interface CourtCasesRepo extends CrudRepository<CourtCase, Integer> {
+public interface CourtCasesRepo extends CrudRepository<CourtCase, Long> {
 
     @Override
     List<CourtCase> findAll();
